✨feat: Dockerfile 및 라우터 수정으로 개발 환경 개선
