+++ app/main.py
@@ -1,8 +1,8 @@
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
-from app.routers.user import router as add_user_favorit_router
 from app.routers.message import router as kakao_router
+from app.routers.user import router as add_user_favorit_router
 
 app = FastAPI(
     title="KakaoTalk Chatbot API",
