+++ tests/unit/routers/test_add_user_favorit.py
@@ -4,9 +4,9 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.sql import text
 
-from app.core.db_manager import db_manager
 from app.main import app
 from app.models import Base, Category, Feature, UserCategory, Users
+from app.utils.db_manager import db_manager
 
 # ✅ 테스트용 SQLite 파일 DB (세션 유지)
 TEST_DATABASE_URL = "sqlite:///./test.db"
