--- tests/unit/routers/test_data/router_cases.json
@@ -1,72 +1,0 @@
-{
-    "handle_message": {
-        "greeting": {
-            "input": {
-                "intent": {},
-                "userRequest": {
-                    "utterance": "안녕하세요"
-                },
-                "bot": {},
-                "action": {}
-            },
-            "expected": {
-                "version": "2.0",
-                "template": {
-                    "outputs": [
-                        {
-                            "simpleText": {
-                                "text": "안녕하세요! 반갑습니다 😊"
-                            }
-                        }
-                    ],
-                    "quickReplies": [
-                        {
-                            "label": "도움말",
-                            "action": "message",
-                            "messageText": "도움말"
-                        },
-                        {
-                            "label": "안녕하세요",
-                            "action": "message",
-                            "messageText": "안녕하세요"
-                        }
-                    ]
-                }
-            }
-        },
-        "help": {
-            "input": {
-                "intent": {},
-                "userRequest": {
-                    "utterance": "도움말"
-                },
-                "bot": {},
-                "action": {}
-            },
-            "expected": {
-                "version": "2.0",
-                "template": {
-                    "outputs": [
-                        {
-                            "simpleText": {
-                                "text": "다음과 같은 기능을 제공합니다:\n1. 인사하기\n2. 도움말 보기"
-                            }
-                        }
-                    ],
-                    "quickReplies": [
-                        {
-                            "label": "도움말",
-                            "action": "message",
-                            "messageText": "도움말"
-                        },
-                        {
-                            "label": "안녕하세요",
-                            "action": "message",
-                            "messageText": "안녕하세요"
-                        }
-                    ]
-                }
-            }
-        }
-    }
-}
