--- tests/unit/utils/response_builder/test_data/response_builder_cases.json
@@ -1,94 +1,0 @@
-{
-    "simple_text": {
-        "input": {
-            "text": "안녕하세요!"
-        },
-        "expected": {
-            "version": "2.0",
-            "template": {
-                "outputs": [
-                    {
-                        "simpleText": {
-                            "text": "안녕하세요!"
-                        }
-                    }
-                ]
-            }
-        }
-    },
-    "basic_card": {
-        "input": {
-            "title": "카드 제목",
-            "description": "카드 설명",
-            "thumbnail_url": "https://example.com/image.jpg"
-        },
-        "expected": {
-            "version": "2.0",
-            "template": {
-                "outputs": [
-                    {
-                        "basicCard": {
-                            "title": "카드 제목",
-                            "description": "카드 설명",
-                            "thumbnail": {
-                                "imageUrl": "https://example.com/image.jpg"
-                            }
-                        }
-                    }
-                ]
-            }
-        }
-    },
-    "quick_replies": {
-        "input": {
-            "response": {
-                "version": "2.0",
-                "template": {
-                    "outputs": [
-                        {
-                            "simpleText": {
-                                "text": "선택해주세요"
-                            }
-                        }
-                    ]
-                }
-            },
-            "quick_replies": [
-                {
-                    "label": "메뉴",
-                    "action": "message",
-                    "messageText": "메뉴 보기"
-                },
-                {
-                    "label": "도움말",
-                    "action": "message",
-                    "messageText": "도움말"
-                }
-            ]
-        },
-        "expected": {
-            "version": "2.0",
-            "template": {
-                "outputs": [
-                    {
-                        "simpleText": {
-                            "text": "선택해주세요"
-                        }
-                    }
-                ],
-                "quickReplies": [
-                    {
-                        "label": "메뉴",
-                        "action": "message",
-                        "messageText": "메뉴 보기"
-                    },
-                    {
-                        "label": "도움말",
-                        "action": "message",
-                        "messageText": "도움말"
-                    }
-                ]
-            }
-        }
-    }
-}
