--- tests/unit/utils/text_utils/test_data/text_utils_cases.json
@@ -1,21 +1,0 @@
-{
-    "extract_keywords": {
-        "input": {
-            "text": "안녕하세요 오늘 날씨가 좋네요"
-        },
-        "expected": ["안녕하세요", "오늘", "날씨가", "좋네요"]
-    },
-    "contains_any_keywords": {
-        "input": {
-            "text": "오늘 날씨가 좋네요",
-            "keywords": ["날씨", "기분", "안녕"]
-        },
-        "expected": true
-    },
-    "create_formatted_list": {
-        "input": {
-            "items": ["사과", "바나나", "오렌지"]
-        },
-        "expected": "1. 사과\n2. 바나나\n3. 오렌지"
-    }
-}
